chore: Refactors battery management, improves SoC & alarms
Refactors the battery management system to enhance State-of-Charge (SoC) accuracy and robustness.

Key changes include:
-   **SoC Calculation:** Introduces dynamic total capacity (`mAh_tot`) with temperature compensation and refined Coulomb counting for charge/discharge tracking (`mAh_n_cycles_charge`, `mAh_neg`). Simplifies overall SoC estimation logic.
-   **Alarm Handling:** Implements distinct alarm types for various battery conditions (e.g., very low battery, anomalous temperature/current/voltage) and integrates reporting into the Modbus interface.
-   **Temperature Measurement:** Updates temperature data types and processing to correctly handle negative temperatures, improving accuracy for NiMH batteries.
-   **Charging Logic:** Adjusts Constant Voltage (CV) charge phase duration, refines charge restart conditions based on capacity, and modifies LED indicators for various charging states.
-   **Non-Volatile Storage:** Updates database parameters to support new capacity tracking and ensures critical data is saved on device shutdown, including before a forced power-off.
-   **Watchdog Integration:** Adds the Watchdog Timer (WWDT) driver to the build system, enabling future system stability features.
-   Refines the current low-pass filter coefficient for smoother readings.

+++ libraries/drivers/BSP/src/bsp_adc.c
@@ -37,6 +37,7 @@
 #include "project_settings.h"
 #include "qpc.h"
 #include "soc_NiMh.h"
+
 /*****************************************************************************
 * Module Preprocessor Constants
 ******************************************************************************/
@@ -103,7 +104,7 @@ bsp_adc_isr_handler(void) {
   lpadc_conv_result_t adc_result;
   uint32_t voltage_mV;
   uint32_t current_mA;
-  uint32_t temp;
+  int16_t temp;
 
   if (LPADC_GetConvResult(LPADC_BASE, &adc_result)) {}
 
@@ -113,7 +114,7 @@ bsp_adc_isr_handler(void) {
   if (LPADC_GetConvResult(LPADC_BASE, &adc_result)) {}
 
   // 16-bit conversion result
-  temp = (uint32_t)bsp_ntc_battery_get_temperature(adc_result.convValue, 65535) * 100U;
+  temp = (int16_t)bsp_ntc_battery_get_temperature(adc_result.convValue, 65535) * 100U;
 
   if (LPADC_GetConvResult(LPADC_BASE, &adc_result)) {}
