+++ libraries/drivers/BSP/src/include/bsp_adc.h
@@ -59,7 +59,7 @@ typedef struct {
     QEvt super;      
     uint32_t vbat;   // mV
     uint32_t ibat;   // mA
-    int32_t tbat;   
+    int16_t tbat;   
     uint32_t soc;
 } AdcInfoEvt;
